pyproject.toml
[tool.poetry]
name = "phat-django"
version = "1.0.0"
description = "My demo Django project for fun"
authors = ["Pham Hong Phat <hphat99@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
django-filter = "^25.1"
posting = "^2.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
requires-python = ">=3.12"
name = "phat-django"
